Local analysis

One question that we often want to ask when considering the dynamics of ecological systems is their response to perturbation. One method to answer such questions is to use local stability analysis where one linearises the system around a fixed point and then studies the dynamics of this linear system to learn about the behaviour of the system as a whole.

MiCRM.jl includes tools for automatically calculating the Jacobian matrix that defines the linearised system and for deriving key stability metrics.

Calculating the Jacobian

MiCRM.jl calculates the Jacobian of a system using ForwardDiff.jl. This automatic-differentiation back end computes derivatives of the ODE system at the terminal state through Analysis.get_jac(sol).

MiCRM.Analysis.get_jac — Function

get_jac(sol)

Numerically calculate jacobian of system from the solution object. Assumes that the system is at equilibrium so the end state is used.

get_jac(sol::DiffEqBase.SteadyStateSolution)

For steady-state solution types.

Dynamic Measures

Once the Jacobian has been calculated, it can be used to derive several stability properties.

MiCRM.Analysis.get_Rins — Function

get_Rins(J, t, w = ones(size(J)[1]))

Caclulate the instantaneous rate of growth of the perturbation u at time t. The observation vector w can be optionally supplied to consider linear combinations of the state variables. This can include any linear combination of state variables such as the mass of specific components of the system as well as total functioning measures.

MiCRM.Analysis.get_stability — Function

get_stability(J)

Determine the stability a system given its jaccobian by testing if the real part of the leading eigenvalue is positive.

MiCRM.Analysis.get_reactivity — Function

get_reactivity(J,u)

Test wether the system is "reactive" to the perturbation u. A reactive is system is one where the initial perturbation is amplified so that the deviation from equilibrium initially increases

MiCRM.Analysis.get_return_rate — Function

get_return_rate(J)

get the rate of return of the system from perturbation. This value is determined by the leading eigenvalue regardless of the direction of the perturbation or the observation vector we choose.